WIP: Fix return values for low covariance matrices
In the case `if numx == 0 or numy == 0:`, the number of returned objects by `compute_ccas` is 6 instead of 5. The change in 139 in should correct this.
Also, in the consequent case `# if x_idxs or y_idxs is all false, return_dict has zero entries` in `get_cca_similarity`, there are several items missing in `return_dict`. The other changes fix this.
Can someone confirm that this is correct?